[WIP] Perform work in macro calling environment
In `define-ffi-function` and `ffi--struct-union-helper` the work to define ffi
functions (using `ffi--prep-cif` and `ffi--dlsym`) and types was done in the
macro environment. This works fine during runtime, but fails for complied code
since the objects would be created at compile time instead of at runtime. By
moving the work to the macro's calling environment we avoid this problem.