[probably not] define-ffi-library: No longer define a function
I think it would be wrong to do this.  I believe that by using a
function we can avoid loading the library until we actually needed.
I might be wrong about this, we should check (certainly before giving
up on the current implementation, but also if we are more inclined to
keep things as is).