define-ffi-library: Locate ffi modules on the load-path
Before this it was not possible to run the tests in an interactive
session.  When given just the name of a library, `ffi--dlopen' tries to
find it on the library path, which already worked for the `test' make
target, because that exported `LD_LIBRARY_PATH', but when running the
tests interactively, that was not the case.

One should only export `LD_LIBRARY_PATH' if absolutely necessary, see
e.g., https://www.hpc.dtu.dk/?page_id=1180.

Use the value of `module-file-suffix' instead of the hardcoded ".so"
suffix, as was done before.  On macOS the suffix may be ".dylib" (the
value of `module-file-suffix' on that platform) or ".so", so we have
to search for both.